+++ src/lib.rs
@@ -9,6 +9,7 @@
 #![cfg_attr(feature = "unstable", feature(test))]
 #[macro_use]
 pub mod counter;
+pub mod account;
 pub mod bank;
 pub mod banking_stage;
 pub mod blob_fetch_stage;
